requirements.txt
discord.py
coc.py>=0.2.1